Bankend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
blinker==1.9.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
crispy-bootstrap5==2025.6
cryptography==46.0.3
cssselect2==0.8.0
defusedxml==0.7.1
distlib==0.4.0
dj-database-url==2.1.0
Django==5.0.1
django-allauth==0.60.1
django-cors-headers==4.3.1
django-crispy-forms==2.5
django-filter==25.2
django-fsm==3.0.1
django-simple-history==3.11.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
filelock==3.20.0
Flask==2.3.3
flask-cors==6.0.2
fonttools==4.61.1
fpdf2==2.8.5
freetype-py==2.5.1
gunicorn==21.2.0
html5lib==1.1
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.3
networkx==3.6.1
numpy==2.4.1
oauthlib==3.3.1
opencv-python==4.8.0.76
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==3.0.0
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycairo==1.29.0
pycparser==2.23
pydyf==0.12.1
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pypdf==6.6.2
pyphen==0.17.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==7.1.0
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
rlPyCairo==0.4.0
scikit-image==0.26.0
scipy==1.17.0
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tifffile==2026.1.14
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.2
vine==5.1.0
virtualenv==20.35.4
wcwidth==0.2.14
weasyprint==68.0
webencodings==0.5.1
Werkzeug==3.1.5
whitenoise==6.6.0
xhtml2pdf==0.2.17
zopfli==0.4.0
